refactor(canvas): split into 13 focused submodules
- Original canvas.rs: 1,647 lines -> new mod.rs: 30 lines (98% reduction)
- Created canvas/ directory with submodules:
  - braille/constants.rs - Braille pattern constants
  - braille/shapes.rs - Shape trait and all shape implementations (504 lines)
  - braille/grid_impl.rs - BrailleGrid implementation (147 lines)
  - braille/context.rs - BrailleContext for drawing (103 lines)
  - braille/mod.rs - Braille module exports (28 lines)
  - transform.rs - 2D transformation matrix (121 lines)
  - clip.rs - ClipRegion for clipping (60 lines)
  - layer.rs - Layer support for composition (90 lines)
  - grid.rs - Grid trait (9 lines)
  - draw.rs - DrawContext for character-based drawing (221 lines)
  - widget.rs - Canvas and BrailleCanvas widgets (102 lines)
  - mod.rs - Main module exports (30 lines)
  - tests.rs - All 24 tests (251 lines)

All tests passing, clippy clean
